+++ src/infos/ecinfos.jl
@@ -19,7 +19,7 @@ export delete_file!, delete_files!, delete_temporary_files!
 export file_description
 export isalphaspin, space4spin, spin4space, flipspin
 export get_options, with_local_options
-export FCIOptions, HCIOptions
+export FCIOptions, CIPHIOptions
 
 include("options.jl")