+++ CHANGELOG.md
@@ -21,7 +21,8 @@
 
 ### Added
 
-* Local options for macros: All calculation macros (`@cc`, `@dfcc`, `@dfhf`, `@dfuhf`, `@dfmcscf`, `@dfmp2`, `@fci`, `@hci`, `@bohf`, `@bouhf`) now accept an optional `begin...end` block to set options locally for that specific call. Options are automatically restored after the call completes. This is the recommended way to set options for individual calculations.
+* FCI and CIPHI (selected CI) methods.
+* Local options for macros: All calculation macros (`@cc`, `@dfcc`, `@dfhf`, `@dfuhf`, `@dfmcscf`, `@dfmp2`, `@fci`, `@ciphi`, `@bohf`, `@bouhf`) now accept an optional `begin...end` block to set options locally for that specific call. Options are automatically restored after the call completes. This is the recommended way to set options for individual calculations.
 * Automatic augmentation of basis sets by additional diffuse or steep functions.
 * Functions to get all elements available in a given basis set (`get_available_elements4basis`) and to output the basis set for a given list of elements in the molpro format (`output_basis`).
 * Augmented basis sets for jkfit vXz-jkfit and def2-universal-jkfit basis sets.
@@ -30,8 +31,8 @@
 * A check for changes of the molecular geometry/basis/fcidump is performed in every macro-command call. If a change is detected, the integrals are set to be recalculated or reloaded from the fcidump file. With this, the user doesn't need to worry about calling `@ECinit` after changing the geometry/basis/fcidump.
 * A test for dummy atoms is added. At the moment, if dummy atoms are detected, the integrals are recalculated. In the future, once we have AO-FDump support, the integrals can be reused.
 * Macros `@loadwf` and `@savewf` to load and save orbitals (etc) from/to trexio dump files. `@copywf` to copy trexio dump files (e.g., to make a local backup).
-* Wavefunction store/start functionality for coupled cluster and selected CI methods. Use `@set wf store="filename.h5"` to store the wavefunction (amplitudes for CC, determinants/coefficients for HCI) to a TREXIO file. Use `@set wf start="filename.h5"` to restart a calculation from a previously stored wavefunction. Multi-state HCI calculations store each state in separate files (e.g., `filename_state2.h5`).
-* `pt2_only` option for HCI calculations (`@set hci pt2_only=true`) to skip variational iterations and compute only the PT2 correction using stored determinants.
+* Wavefunction store/start functionality for coupled cluster and selected CI methods. Use `@set wf store="filename.h5"` to store the wavefunction (amplitudes for CC, determinants/coefficients for CIPHI) to a TREXIO file. Use `@set wf start="filename.h5"` to restart a calculation from a previously stored wavefunction. Multi-state CIPHI calculations store each state in separate files (e.g., `filename_state2.h5`).
+* `pt2_only` option for CIPHI calculations (`@set ciphi pt2_only=true`) to skip variational iterations and compute only the PT2 correction using stored determinants.
 
 ### Fixed